requirements.txt
streamlit>=1.31.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
pandas>=2.2.0
plotly==5.17.0
fpdf==1.7.2
python-dateutil==2.8.2
bcrypt>=4.0.0
python-docx==1.2.0
